Fixed bug in log.py
I could have sworn I submitted this last night, but now I don't see a pull request for it.  Apologies if this is a duplicate.

In the send_mail_msg function, the call to MIMEText requires an object with an encode attribute (presumably a string).  It won't convert an object to a string for you.