The broker account value is changed here, not the total capital. This is curious, because total capital is what is allocated to strategies, and thus what position sizes are based on, so position sizes will not be adjusted to reflect the deposit or withdrawal.
That seems like it would fine if using "fixed" capital, or "half" capital accumulation, and only profits above the maximum capital are withdrawn (which I think is Rob's use-case).  But it doesn't work at all with "full" capital - a manual adjustment to total capital would also be required in that case.  Maybe that is by design?

Might need to revisit this in the future.